pyproject.toml
[project]
name = "mcp-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.4",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
mcp-service = "mcp_service.server:main"

